added CSV export to Statistics.cs
