+++ PoGo.NecroBot.Logic/Utils/Statistics.cs
@@ -3,10 +3,12 @@
 #region using directives
 
 using System;
+using System.IO;
 using System.Globalization;
+using System.Collections.Generic;
 using System.Linq;
 using POGOProtos.Networking.Responses;
-
+using PoGo.NecroBot.Logic.PoGoUtils;
 #endregion
 
 // ReSharper disable CyclomaticComplexity
@@ -33,7 +35,7 @@ namespace PoGo.NecroBot.Logic.Utils
         {
             _currentLevelInfos = GetCurrentInfo(inventory);
             DirtyEvent?.Invoke();
-            inventory.ExportPokemonToCSV($"{_playerName}.csv", ToString());
+            ExportPokemonToCSV(inventory, $"{_playerName}.csv", ToString());
         }
 
         public event StatisticsDirtyDelegate DirtyEvent;
@@ -97,6 +99,77 @@ namespace PoGo.NecroBot.Logic.Utils
         {
             return
                 $"{_playerName} - Runtime {FormatRuntime()} - Lvl: {_currentLevelInfos} | EXP/H: {TotalExperience/GetRuntime():0} | P/H: {TotalPokemons/GetRuntime():0} | Stardust: {TotalStardust:0} | Transfered: {TotalPokemonsTransfered:0} | Items Recycled: {TotalItemsRemoved:0}";
+        }
+        public async void ExportPokemonToCSV(Inventory inventory, string filename , string playerstats = "")
+        {
+            string export_path = Path.Combine("Exports");
+            if (!Directory.Exists(export_path))
+                Directory.CreateDirectory(export_path);
+            if (Directory.Exists(export_path))
+            {
+                try
+                {
+                    string pokelist_file = Path.Combine(export_path, filename);
+                    if (File.Exists(pokelist_file))
+                        File.Delete(pokelist_file);
+                    string ls = System.Globalization.CultureInfo.CurrentCulture.TextInfo.ListSeparator;
+                    File.WriteAllText(pokelist_file, $"{playerstats.Replace(",", $"{ls}")}");
+                    string header = "PokemonID,Name,NickName,CP,MaxCP,Perfection,Attack 1,Attack 2,HP,Attk,Def,Stamina,Familie Candies";
+                    using (var w = File.AppendText(pokelist_file))
+                    {
+                        w.WriteLine("");
+                        w.WriteLine($"{ $"{header.Replace(",", $"{ls}")}"}");
+                        
+                    }
+                    
+                    var AllPokemonInInv = await inventory.GetPokemons();
+                    AllPokemonInInv = AllPokemonInInv.Where(p => p.DeployedFortId == string.Empty).OrderByDescending(p => p.PokemonId);
+                    var myPokemonSettings = await inventory.GetPokemonSettings();
+                    var pokemonSettings = myPokemonSettings.ToList();
+                    var myPokemonFamilies = await inventory.GetPokemonFamilies();
+                    var pokemonFamilies = myPokemonFamilies.ToArray();
+
+                    using (var w = File.AppendText(pokelist_file))
+                    {
+                        w.WriteLine("");
+                        foreach (var pokemon in AllPokemonInInv)
+                        {
+                            var settings = pokemonSettings.Single(x => x.PokemonId == pokemon.PokemonId);
+                            var familiecandies = pokemonFamilies.Single(x => settings.FamilyId == x.FamilyId).Candy;
+                            string perfection = $"{ PokemonInfo.CalculatePokemonPerfection(pokemon).ToString("0.00") }%";
+                            perfection = perfection.Replace(",", ls == "," ? "." : ",");
+                            string content_part1 = $"{(int)pokemon.PokemonId},{pokemon.PokemonId},{pokemon.Nickname},{pokemon.Cp},{PokemonInfo.CalculateMaxCp(pokemon).ToString().PadLeft(4, ' ')},";
+                            string content_part2 = $",{pokemon.Move1},{pokemon.Move2},{pokemon.Stamina},{pokemon.IndividualAttack},{pokemon.IndividualDefense},{pokemon.IndividualStamina},{familiecandies}";
+                            string content = $"{content_part1.Replace(",", $"{ls}")}{perfection}{content_part2.Replace(",", $"{ls}")}";
+                            w.WriteLine($"{content}");
+
+                        }
+                        w.WriteLine("");
+                    }
+                    using (var w = File.AppendText(pokelist_file))
+                    {
+                        List<int> allnums = new List<int>();
+                        for (int i = 1; i <= 151; i++)
+                            allnums.Add(i);
+                        foreach (var pokemon in AllPokemonInInv)
+                        {
+                            if  (allnums.Contains((int)pokemon.PokemonId))
+                            {
+                                allnums.Remove((int)pokemon.PokemonId);
+                            }
+                        }
+                        w.WriteLine($"Catched: {151 - allnums.Count}");
+                        w.WriteLine("Missing:");
+                        foreach (int num in allnums)
+                        w.WriteLine($"{num}");
+                        w.Close();
+                    }
+                }
+                catch
+                {
+
+                }
+            }
         }
     }
 }
